feat: integrate runtime, SQL media, and dashboard updates (#19)
* feat: integrate runtime, SQL media, and dashboard updates


* fix: require a secure Pillow baseline


* fix: require a secure pypdf baseline


* fix: verify fsspec license provenance


* docs: refresh evaluation translations


* test: validate reviewed license overrides


---------

+++ src/ray/klein/runtime/graph/vertex_spec.py
@@ -33,7 +33,7 @@ class VertexSpec:
         return self.operator.runtime_info.batch_size
 
     @property
-    def batch_timeout(self) -> int | None:
+    def batch_timeout(self) -> float | None:
         return self.operator.runtime_info.batch_timeout
 
     @property
